Added corrupted fractured test, fixed issues uncovered.
Added test intended to uncover the order of the "Fractured" and
"Corrupted" final tags on items. Fixed that issue, but uncovered an
issue related to item properties. Added elemental damage is reflected in
item properties by default, but it's the only property (so far) that has
multiple values in a list. Previously, secondary and tertiary values
were ignored, which lead to weapons with multiple elemental damages only
listing the first one.