Added function to generate whisper text.
The most common way for people to buy items from other players on
POE is to search the public stash tabs with a trade sie, then have the
trade site generate a "whisper" message that the user then pastes into
the game client. There are additional tools which can parse incoming
trade whispers and display overlay graphics to make finding the relevant
items easier for the seller. This seems like a good hack for finding
items.