Reworked item->str to be more flexible, fixing prophecies.
item->str has always been an absolute mess, so I cleaned it up by codifying the
concept of mods and blocks into a spec, and coding based on those types with
post conditions. This made the code much more modular, and allowed dealing with
prophecies cleanly. Prophecies bizarrely invert the order of some of the
descriptions, so the new coding technique makes it much easier to efficiently
handle exceptions like that.